+++ Changelog.md
@@ -55,6 +55,7 @@
 - Fix: Anyone can convert a Frame/Reinforced Iron Trapdoor to a Keypad/Passcode-protected Trapdoor
 - Fix: The Keypad Trapdoor's signal length option cannot be set to 0 despite the description saying it can
 - Removed: Ability to change the subpage in the SecurityCraft Manual using the arrow keys. This is still possible by scrolling while holding down CTRL
+- Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
 
 --------------------------Changelog for v1.9.8-beta2 of SecurityCraft--------------------------